fix: split test_test_filter.py to satisfy 1000-line size limit
Move conftest filter plugin tests (TestConftestFilterPlugin) and shadow-diff
tests (TestShadowDiff) into new test_test_filter_plugin.py. The file grew to
1098 lines after adding TestReexportClosureIntegration, breaching the groupE
1000-line limit enforced by test_test_suite_oversized_files_split.
